Move HAMT tests to 'lib' or '_unautomated'
These are not per se testing the runtime itself. Rather, they are checking that
the standard library actually is correct.

Furthermore, the 'hamt_stress.trp' and 'hamt_perf.trp' tests include timing the
time to run certain operations. These break on other machines (and presumably
also on the same machine).