Add modules to compiler (WIP)
The compiler is extended with two flags:
- One to compile a troupe program as a Module. This outputs the IR
  and also its (SHA256) hash.
- Another to include directory for modules. This is needed to find
  the hash value of any module that was required.

Throughout the frontend, the modules are mainly passed along at the
program level without further treatment.

At the IR level and beyond (backend), the module dependencies are
moved from the program level to the function definitions. Since
functions can be sent independently to others, this allows the
different functions to keep track of their respective dependencies.