Replace library imports with new local module system
This only applies to the standard library in the 'lib/' folder;
user-defined libraries that are opened via the `require` instead of
`import` statement are ignored.

What this does provide is a clean-up of the previous hacky library
system. Doing so, we fix the inability to compute constants as part of
library initialisation (issue #55).

This provides the foundation for us to build the nomadic module
system.