feat: add StdIO module & clean up IO in TCB
This removes some derivable operations out of the TCB, e.g. `fprintln`
and `fprintlnWithLabels`. At the same time, `inputLine` has been
semi-generalised to be `freadln` (read a line from a file) and `stdin`
is now a derivable capability similar to the one needed for printing
to `stdout`. This also fixes the missing access to `stderr`.

For now, there are lots of `runtime/core` tests that rely on the
preamble functions, `print` and so on. In the long run, these tests
should be rewritten such that they are completely independent.

The tests for the preamble itself have been moved into a separate
folder.