docs: add TODO on equality checking for Raw operations
This operation seems only designed for the values of the Raw type as
is now defined in rawUtil.mts. Yet, it is hard to get it into the type
system, as the basic types (number, string, and boolean) do not have a
'_troupeType' value.

Either this algorithm needs some rewriting and/or we need to redesign
the type definition a bit to make it easier for TypeScript to follow.