feat: add `sha256` builtin
For the addition of nomadic values, we need to be able to hash these
values. For me to be able to test it, I would like to have it as a
builtin to see it work.

A Troupe programmer probably also want access to a cryptographic hash
function of their values (unlike the algorithmic hash in the `Hash`
module).