refactor: rename Thread::raise* function
The notion of a 'currentThread' should only be part of the Scheduler,
not the individual thread. Furthermore, the thread itself already
provides the context of raising a thread's level; there is no reason
to say it twice.

This also spells out 'PC' as 'ProgramCounter' and 'BlockingLev' as
'BlockingLevel'; the previous naming scheme was inconsistent (and also
'lvl' ought to be the shortening of 'level', not 'lev')